Rename cron job index.js
